refactor: move User struct to models/user.go
